+++ include/solar/entry.hpp
@@ -1,3 +1,4 @@
 #pragma once
 
 #include "solar/entry/profile.hpp"
+#include "solar/entry/zephyr.hpp"