+++ include/solar/rtos/config.hpp
@@ -2,7 +2,7 @@
 
 #include <cstdint>
 
-#include "low_level/rtos/config.hpp"
+#include <zephyr/kernel.h>
 
 namespace solar::rtos
 {
@@ -11,15 +11,15 @@ constexpr std::uint8_t PriorityStepsPerBand = 6;
 constexpr std::uint8_t PriorityBands = 5;
 constexpr std::uint8_t PriorityLevels = PriorityStepsPerBand * PriorityBands;
 
-using NativePriority = low_level::rtos::NativePriority;
+using NativePriority = int;
 
 constexpr NativePriority max_native_priority()
 {
-    return low_level::rtos::max_native_priority();
+    return CONFIG_NUM_PREEMPT_PRIORITIES > 0 ? CONFIG_NUM_PREEMPT_PRIORITIES : 1;
 }
 
 } // namespace solar::rtos
 
 #ifndef SOLAR_RTOS_ASSERT
-#define SOLAR_RTOS_ASSERT(expr) LOW_LEVEL_RTOS_ASSERT(expr)
+#define SOLAR_RTOS_ASSERT(expr) __ASSERT_NO_MSG(expr)
 #endif