+++ include/solar/rtos/critical_section.hpp
@@ -1,11 +1,21 @@
 #pragma once
 
-#include "low_level/rtos/sync.hpp"
+#include <zephyr/irq.h>
 
 namespace solar::rtos
 {
 
-using CriticalSection = low_level::rtos::CriticalSection;
+class CriticalSection
+{
+public:
+    CriticalSection() : key_(irq_lock()) {}
+    ~CriticalSection() { irq_unlock(key_); }
+
+    CriticalSection(const CriticalSection &) = delete;
+    CriticalSection &operator=(const CriticalSection &) = delete;
+
+private:
+    unsigned int key_{};
+};
 
 } // namespace solar::rtos
-