+++ include/solar/rtos/scheduler.hpp
@@ -1,21 +1,23 @@
 #pragma once
 
-#include "low_level/rtos/scheduler.hpp"
-
 namespace solar::rtos
 {
 
-using SchedulerState = low_level::rtos::SchedulerState;
+enum class SchedulerState
+{
+    NotStarted,
+    Running,
+    Suspended,
+};
 
 inline void start_scheduler()
 {
-    low_level::rtos::start_scheduler();
+    // Zephyr starts the scheduler before application main.
 }
 
 inline SchedulerState scheduler_state()
 {
-    return low_level::rtos::scheduler_state();
+    return SchedulerState::Running;
 }
 
 } // namespace solar::rtos
-