+++ include/solar/rtos/time.hpp
@@ -2,14 +2,15 @@
 
 #include <chrono>
 
+#include <zephyr/kernel.h>
+
 #include "solar/core/time.hpp"
-#include "low_level/rtos/time.hpp"
 
 namespace solar::rtos
 {
 
-using Tick = low_level::rtos::Tick;
-inline constexpr Tick WaitForever = low_level::rtos::WaitForever;
+using Tick = std::int64_t;
+inline constexpr Tick WaitForever = -1;
 
 using Milliseconds = solar::Milliseconds;
 using Microseconds = solar::Microseconds;
@@ -20,23 +21,36 @@ using Seconds = solar::Seconds;
  */
 constexpr Tick to_ticks(Tick ticks)
 {
-    return low_level::rtos::to_ticks(ticks);
+    return ticks;
 }
 
 template <class Rep, class Period>
 constexpr Tick to_ticks(std::chrono::duration<Rep, Period> duration)
 {
-    return low_level::rtos::to_ticks(duration);
+    return static_cast<Tick>(std::chrono::duration_cast<Milliseconds>(duration).count());
 }
 
 constexpr Milliseconds to_milliseconds(Tick ticks)
 {
-    return low_level::rtos::to_milliseconds(ticks);
+    return Milliseconds{ticks < 0 ? 0 : ticks};
+}
+
+inline k_timeout_t to_timeout(Tick ticks)
+{
+    if (ticks == WaitForever)
+    {
+        return K_FOREVER;
+    }
+    if (ticks <= 0)
+    {
+        return K_NO_WAIT;
+    }
+    return K_MSEC(ticks);
 }
 
 inline Tick now_ticks()
 {
-    return low_level::rtos::now_ticks();
+    return static_cast<Tick>(k_uptime_get());
 }
 
 inline Milliseconds now()